backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.3
python-multipart==0.0.9
pydantic==2.9.2
PyJWT==2.9.0
bcrypt==4.2.0